+++ sembast/lib/src/record_ref_impl.dart
@@ -109,6 +109,10 @@ extension SembastRecordRefExtension<K, V> on RecordRef<K, V> {
   Future<V?> get(DatabaseClient databaseClient) async =>
       (await getSnapshot(databaseClient))?.value;
 
+  /// Get a record value from the database synchronously.
+  V? getSync(DatabaseClient databaseClient) =>
+      getSnapshotSync(databaseClient)?.value;
+
   /// Get a record snapshot from the database.
   Future<RecordSnapshot<K, V>?> getSnapshot(
       DatabaseClient databaseClient) async {
@@ -116,8 +120,17 @@ extension SembastRecordRefExtension<K, V> on RecordRef<K, V> {
 
     var record = await client
         .getSembastStore(store)
-        .txnGetRecord(client.sembastTransaction, key as Key);
-    return record?.cast<K, V>();
+        .txnGetRecordSnapshot<K, V>(client.sembastTransaction, key);
+    return record;
+  }
+
+  /// Get a record snapshot from the database synchronously.
+  RecordSnapshot<K, V>? getSnapshotSync(DatabaseClient databaseClient) {
+    var client = getClient(databaseClient);
+
+    return client
+        .getSembastStore(store)
+        .txnGetRecordSnapshotSync<K, V>(client.sembastTransaction, key);
   }
 
   /// Get a stream of a record snapshot from the database.
@@ -153,6 +166,14 @@ extension SembastRecordRefExtension<K, V> on RecordRef<K, V> {
     return client
         .getSembastStore(store)
         .txnRecordExists(client.sembastTransaction, key as Key);
+  }
+
+  /// Return true if the record exists synchronously.
+  bool existsSync(DatabaseClient databaseClient) {
+    var client = getClient(databaseClient);
+    return client
+        .getSembastStore(store)
+        .txnRecordExistsSync(client.sembastTransaction, key as Key);
   }
 
   /// Delete the record. Returns the key if deleted, null if not found.